requirements.txt
asciitree==0.3.3
av==11.0.0
certifi==2023.11.17
charset-normalizer==3.3.2
cmake==3.28.1
contourpy==1.2.0
cycler==0.12.1
easydict==1.11
fasteners==0.19
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
hiera-transformer==0.1.2
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
lit==17.0.6
MarkupSafe==2.1.3
matplotlib==3.8.0
mpmath==1.3.0
networkx==3.2.1
numcodecs==0.12.1
numpy==1.26.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
psutil==5.9.7
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.1
scikit-learn==1.3.1
scipy==1.11.4
six==1.16.0
sympy==1.12
threadpoolctl==3.2.0
timm==0.9.12
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.1
triton==2.0.0
turbojpeg==0.0.2
typing_extensions==4.9.0
urllib3==2.1.0
zarr==2.16.1
